Improving running over signal MC

+++ python/run_BHistograms.py
@@ -87,31 +87,58 @@ def RunBHistogramsSignal(analysis, sample, files_per_job=1, retar=False, data_so
 	os.system("mkdir -pv " + working_directory)
 	os.chdir(working_directory)
 
-	command = "condor_cmsRun"
-	if retar:
-		command += " --retar "
-	#input_txt = open("tmp.txt", 'w')
-	#input_txt.write(analysis_config.files_QCDBEventTree[sample] + "\n")
-	#input_txt.close()
-	command += " --file-list=" + analysis_config.files_QCDBEventTree[sample] + " "
-	command += " --files-per-job=" + str(files_per_job)
-	command += " --submit-file=submit_" + analysis + "_" + sample + ".jdl "
-	#command += " --output-file=" + output_prefix + "_" + sample + ".root "
-	command += " --output-tag=BHistograms_" + sample + " "
-	command += " --run "
-	command += "  " + dijet_directory + "/CMSSW_5_3_32_patch3/src/MyTools/RootUtils/scripts/cmsRun_wrapper.sh " + analysis_config.analysis_cfgs[analysis] 
-	command += " dataSource=simulation "
-	command += " dataType=signal "
-	command += " signalMass=" + str(analysis_config.simulation.signal_sample_masses[sample]) + " "
-	#command += "inputFiles=" + os.path.basename(input_files[sample])
-	output_filename = os.path.basename(analysis_config.get_b_histogram_filename(analysis, sample)).replace(".root", "_\$\(Cluster\)_\$\(Process\).root")
-	command += " outputFile=" + os.path.basename(analysis_config.get_b_histogram_filename(analysis, sample)).replace(".root", "_\$\(Cluster\)_\$\(Process\).root")
-	print command
-	os.system(command)
-	os.system("rm -f tmp.txt")
+	method = "csub"
+	if method == "csub":
+		input_files_txt = open(analysis_config.files_QCDBEventTree[sample], 'r')
+		input_files = [line.strip() for line in input_files_txt]
+		input_files_txt.close()
+
+		bash_script_path = working_directory + "/run_" + analysis + "_" + sample + ".sh"
+		bash_script = open(bash_script_path, 'w')
+		bash_script.write("#!/bin/bash\n")
+		bash_script.write("input_files=( " + " ".join([os.path.basename(x) for x in input_files]) + " )\n")
+		output_filename = os.path.basename(analysis_config.get_b_histogram_filename(analysis, sample)).replace(".root", "_$1.root")
+		bash_script.write("cmsRun " 
+			+ os.path.basename(analysis_config.analysis_cfgs[analysis]) 
+			+ " dataSource=simulation"
+			+ " dataType=signal"
+			+ " signalMass=" + str(analysis_config.simulation.signal_sample_masses[sample])
+			+ " outputFile=" + output_filename
+			+ " inputFiles=file:${input_files[$1]}\n"
+		)
+		bash_script.close()
+
+		submit_command = "csub " + bash_script_path + " --cmssw "
+		if not retar:
+			submit_command += " --no_retar"
+		submit_command += " -F " + ",".join(input_files) + "," + analysis_config.analysis_cfgs[analysis] + " -n " + str(len(input_files))
+		os.system(submit_command)
+	else:
+		command = "condor_cmsRun"
+		if retar:
+			command += " --retar "
+		#input_txt = open("tmp.txt", 'w')
+		#input_txt.write(analysis_config.files_QCDBEventTree[sample] + "\n")
+		#input_txt.close()
+		command += " --file-list=" + analysis_config.files_QCDBEventTree[sample] + " "
+		command += " --files-per-job=" + str(files_per_job)
+		command += " --submit-file=submit_" + analysis + "_" + sample + ".jdl "
+		#command += " --output-file=" + output_prefix + "_" + sample + ".root "
+		command += " --output-tag=BHistograms_" + sample + " "
+		command += " --run "
+		command += "  " + dijet_directory + "/CMSSW_5_3_32_patch3/src/MyTools/RootUtils/scripts/cmsRun_wrapper.sh " + analysis_config.analysis_cfgs[analysis] 
+		command += " dataSource=simulation "
+		command += " dataType=signal "
+		command += " signalMass=" + str(analysis_config.simulation.signal_sample_masses[sample]) + " "
+		#command += "inputFiles=" + os.path.basename(input_files[sample])
+		output_filename = os.path.basename(analysis_config.get_b_histogram_filename(analysis, sample)).replace(".root", "_\$\(Cluster\)_\$\(Process\).root")
+		command += " outputFile=" + os.path.basename(analysis_config.get_b_histogram_filename(analysis, sample)).replace(".root", "_\$\(Cluster\)_\$\(Process\).root")
+		print command
+		os.system(command)
+		os.system("rm -f tmp.txt")
 	postprocessing_file = open('postprocessing_' + analysis + "_" + sample + ".sh", 'w')
 	postprocessing_file.write("#!/bin/bash\n")
-	postprocessing_file.write("hadd " + working_directory + "/" + os.path.basename(analysis_config.get_b_histogram_filename(analysis, sample)) + " " + output_filename.replace("_\$\(Cluster\)_\$\(Process\)", "*") + "\n")
+	postprocessing_file.write("hadd " + working_directory + "/" + os.path.basename(analysis_config.get_b_histogram_filename(analysis, sample)) + " " + output_filename.replace("_\$\(Cluster\)_\$\(Process\)", "*").replace("$1", "*") + "\n")
 	postprocessing_file.close()
 
 	# cd back
